pyproject.toml
[tool.poetry]
name = "habits"
version = "0.1.0"
description = "Сервис для трекинга привычек с напоминаниями в Telegram"
authors = ["Ваше Имя <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.14"
celery = "^5.5"
redis = "^4.5"
django-cors-headers = "^4.1"
drf-yasg = "^1.21"
requests = "^2.31"
python-dotenv = "^1.0"
dj-database-url = "^3.0.0"
django-celery-beat = "^2.8.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"

[tool.poetry.dev-dependencies]
pytest = "^7.3"
flake8 = "^6.0"
black = "^24.1"
isort = "^5.12"

[tool.poetry.scripts]
manage = "project.manage:main"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

